code cleanup
- moved stubs from the test class, to keep it more readable.
